Works under cygwin and linux

+++ utils/recycle.c
@@ -84,7 +84,7 @@ remalloc (len, purpose)
   char *x = (char *) malloc (len);
 
   if (!x) {
-    fprintf (stderr, "malloc of %lu failed for %s\n", len, purpose);
+    fprintf (stderr, "malloc of %lu failed for %s\n", (unsigned long) len, purpose);
     exit (SUCCESS);
   }
   return x;
